adding reading from a json file and created a test folder
