+++ server/api-manager.js
@@ -20,7 +20,7 @@ const SITEID = "-99";
 
 import MindbodyRequest from "./requests.js";
 import QueryString from "query-string";
-export default class MindbodyQueries {
+class MindbodyQueries {
   constructor() {
     this.requestNum = 0;
     this.authToken = null;
